requirements.txt
asgiref==3.5.0
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
idna==3.3
mypy-extensions==0.4.3
ndg-httpsclient==0.5.1
pathspec==0.9.0
platformdirs==2.5.1
psycopg2==2.9.3
pusher==3.2.0
pyasn1==0.4.8
pycparser==2.21
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
